Preprocessing_PIPELINE/requirements.txt
spacy
nltk
flair
pyap
phonenumbers
email-validator
vaderSentiment
sentence-transformers
fuzzywuzzy
requests
beautifulsoup4
selenium
fitz
pytesseract
pdf2image
numpy
faiss-cpu
scikit-learn
matplotlib
seaborn
tkinter